requirements.txt
pandas
requests
mysql-connector-python
scikit-learn
joblib
python-dotenv
